To Do List / week 7
In this application, the things you want to keep in the form of a list are entered into the system and appear as a list. Items in the list can be edited and deleted later. I used some icons in the application. These icons were achieved by installing fontawesome packages.